Renamed files with 'astronomy' prefix.
